+++ packages/react-app/package.json
@@ -35,6 +35,7 @@
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-icons": "^4.3.1",
+    "react-router-dom": "6",
     "react-scripts": "4.0.3",
     "styled-components": "^5.3.3",
     "typechain": "^8.0.0"
