+++ packages/react-app/src/App.tsx
@@ -1,65 +1,19 @@
-import { shortenAddress, useEthers, useLookupAddress } from "@usedapp/core";
-import { useEffect, useState } from "react";
-import { Body, Container, Header } from "./components";
-import { Mint } from './components/mint'
-import { FetchData } from './components/fetch'
-import { Button } from '@chakra-ui/react'
-
-function WalletButton() {
-  const [rendered, setRendered] = useState("");
-
-  const ens = useLookupAddress();
-  const { account, activateBrowserWallet, deactivate, error } = useEthers();
-
-  useEffect(() => {
-    if (ens) {
-      setRendered(ens);
-    } else if (account) {
-      setRendered(shortenAddress(account));
-    } else {
-      setRendered("");
-    }
-  }, [account, ens, setRendered]);
-
-  useEffect(() => {
-    if (error) {
-      console.error("Error while connecting wallet:", error.message);
-    }
-  }, [error]);
-
-  return (
-    <Button
-    
-      onClick={() => {
-        if (!account) {
-          activateBrowserWallet();
-        } else {
-          deactivate();
-        }
-      }}
-      colorScheme='purple'
-      margin= '4'
-      size='sm'
-      variant='outline'
-      >
-      {rendered === "" && "Connect Wallet"}
-      {rendered !== "" && rendered}
-    </Button>
-  );
-}
+import { Home } from './pages/home'
+import { Nft } from './pages/nft'
+import { NotFound } from './pages/notFound'
+import { Routes, Route } from "react-router-dom";
 
 function App() {
 
   return (
-    <Container>
-      <Header>
-        <WalletButton />
-      </Header>
-      <Body>
-        <Mint />
-        <FetchData />
-      </Body>
-    </Container>
+
+    <Routes>
+      <Route path="/" element={<Home />} />
+      <Route path=':address/:id' element={<Nft />} />
+      <Route path="*" element={<NotFound />} />
+
+    </Routes>
+
   );
 }
