+++ packages/react-app/src/gen/types/Erc721.ts
